+++ src/component/table/table.sass
@@ -2,6 +2,7 @@
   margin: 20px auto
   max-width: 800px
   border-radius: 8px
+  background-color: #fff
 
 .table-header
   font-weight: bold
